+++ test/recording_studio_terms_and_conditions_test.rb
@@ -222,6 +222,7 @@ class RecordingStudioTermsAndConditionsTest < Minitest::Test
     assert_includes layout, 'stylesheet_link_tag "tailwind"'
     assert_includes layout, '<html data-theme="rounded">'
     assert_includes layout, "dummy_top_nav"
+    assert_includes layout, "content_for?(:skip_top_nav)"
     assert_includes layout, "anchor_href"
     refute_includes layout, "page_nav_options[:back_url]"
     refute_includes layout, "page_nav_options[:anchor_url]"
@@ -346,6 +347,9 @@ class RecordingStudioTermsAndConditionsTest < Minitest::Test
     refute_includes admin_new, "card.footer"
     assert_includes admin_new, 'class="inline-block"'
     assert_includes admin_new, "terms_admin_hub_path"
+    assert_includes admin_new, 'content_for :skip_top_nav'
+    assert_includes admin_new, "New Terms and Conditions"
+    refute_includes admin_new, "Write terms"
     refute_includes engine_source("#{views}/admin/terms/edit.html.erb"), "FlatPack::Card::Component"
     assert_includes engine_source("#{views}/admin/terms/edit.html.erb"), 'class="inline-block"'
     refute_includes admin_show, "FlatPack::Card::Component"